Simplify prometheus rules
